Check if file present
This is good for people who want to make sure a file is there and alert if the file is no there. For my case this is good for Nginx matchers. We have several matchers for our micro services and we want to loop through them to confirm they are there.